Move getUserId to context, fix TypeMap import
