+++ src/prisma.ts
@@ -1,0 +1,6 @@
+import { Prisma } from './prisma-client'
+
+export const prisma = new Prisma({
+  endpoint: process.env.PRISMA_ENDPOINT,
+  secret: process.env.PRISMA_SECRET,
+})
