+++ src/resolvers/AuthPayload.ts
@@ -1,5 +1,5 @@
 import { AuthPayloadResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 import { UserParent } from './User'
 
 export interface AuthPayloadParent {
