+++ src/resolvers/City.ts
@@ -1,5 +1,5 @@
 import { CityResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 
 export interface CityParent {
   id: string
