+++ src/resolvers/CreditCardInformation.ts
@@ -1,5 +1,5 @@
 import { CreditCardInformationResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 import { PaymentAccountParent } from './PaymentAccount'
 
 export interface CreditCardInformationParent {
