+++ src/resolvers/ExperiencesByCity.ts
@@ -1,5 +1,5 @@
 import { ExperiencesByCityResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 import { ExperienceParent } from './Experience'
 import { CityParent } from './City'
