+++ src/resolvers/GuestRequirements.ts
@@ -1,5 +1,5 @@
 import { GuestRequirementsResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 
 export interface GuestRequirementsParent {
   govIssuedId: boolean
