+++ src/resolvers/Location.ts
@@ -1,5 +1,5 @@
 import { LocationResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 
 export interface LocationParent {
   id: string
