+++ src/resolvers/Neighbourhood.ts
@@ -1,5 +1,5 @@
 import { NeighbourhoodResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 import { PictureParent } from './Picture'
 import { CityParent } from './City'
