+++ src/resolvers/Notification.ts
@@ -1,5 +1,5 @@
 import { NotificationResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 import { UserParent } from './User'
 
 export type NOTIFICATION_TYPE =
