+++ src/resolvers/PaymentAccount.ts
@@ -1,5 +1,5 @@
 import { PaymentAccountResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 import { UserParent } from './User'
 import { PaymentParent } from './Payment'
 import { PaypalInformationParent } from './PaypalInformation'
