+++ src/resolvers/Policies.ts
@@ -1,5 +1,5 @@
 import { PoliciesResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 
 export interface PoliciesParent {
   checkInEndTime: number
