+++ src/resolvers/Query.ts
@@ -1,6 +1,5 @@
-import { getUserId } from '../utils'
 import { QueryResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 
 export interface QueryParent {}
 
@@ -19,7 +18,8 @@ export const Query: QueryResolvers.Type<TypeMap> = {
       }
     })
   },
-  topHomes: (_parent, _args, ctx) => ctx.db.places({ orderBy: 'popularity_DESC' }),
+  topHomes: (_parent, _args, ctx) =>
+    ctx.db.places({ orderBy: 'popularity_DESC' }),
   homesInPriceRange: (_parent, { min, max }, ctx) =>
     ctx.db.places({
       where: {
@@ -56,7 +56,7 @@ export const Query: QueryResolvers.Type<TypeMap> = {
     bookings: null,
   }),
   myLocation: async (_parent, _args, ctx) => {
-    const id = getUserId(ctx)
+    const id = ctx.getUserId()
 
     const locations = await ctx.db.locations({
       where: {
