+++ src/resolvers/Review.ts
@@ -1,5 +1,5 @@
 import { ReviewResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 
 export interface ReviewParent {
   accuracy: number
