+++ src/resolvers/Subscription.ts
@@ -1,5 +1,5 @@
 import { SubscriptionResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 
 export interface SubscriptionParent {}
