+++ src/resolvers/Viewer.ts
@@ -1,8 +1,7 @@
 import { ViewerResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 import { UserParent } from './User'
 import { BookingParent } from './Booking'
-import { getUserId } from '../utils'
 
 export interface ViewerParent {
   me: UserParent
@@ -11,12 +10,12 @@ export interface ViewerParent {
 
 export const Viewer: ViewerResolvers.Type<TypeMap> = {
   me: (_parent, _args, ctx) => {
-    const id = getUserId(ctx)
+    const id = ctx.getUserId()
 
     return ctx.db.user({ id })
   },
   bookings: async (_parent, _args, ctx) => {
-    const id = getUserId(ctx)
+    const id = ctx.getUserId()
     const bookings =
       (await ctx.db.bookings({ where: { bookee: { id } } })) || []
