+++ src/resolvers/index.ts
@@ -1,5 +1,5 @@
 import { IResolvers } from '../generated/resolvers'
-import { TypeMap } from '../types/TypeMap'
+import { TypeMap } from './types/TypeMap'
 
 import { Query } from './Query'
 import { Mutation } from './Mutation'
