--- src/types/Context.ts
@@ -1,5 +1,0 @@
-import { Prisma } from '../prisma-client'
-
-export interface Context {
-  db: Prisma
-}
