--- src/utils.ts
@@ -1,25 +1,0 @@
-import * as jwt from 'jsonwebtoken'
-import { Prisma } from './prisma-client'
-
-export interface Context {
-  db: Prisma
-}
-
-export function getUserId(context) {
-  const Authorization = context.request.get('Authorization')
-  if (Authorization) {
-    const token = Authorization.replace('Bearer ', '')
-    const { userId } = jwt.verify(token, process.env.APP_SECRET!) as {
-      userId: string
-    }
-    return userId
-  }
-
-  throw new AuthError()
-}
-
-export class AuthError extends Error {
-  constructor() {
-    super('Not authorized')
-  }
-}
